+++ Modules/_testclinic.c
@@ -6,6 +6,7 @@
 #undef NDEBUG
 
 #include "Python.h"
+#include "pycore_object.h"        // _PyObject_IsFreed()
 
 
 // Used for clone_with_conv_f1 and clone_with_conv_v2
